% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hathidy.R
\name{add_chunks}
\alias{add_chunks}
\title{Add a set of 'chunks' to a tibble with page-level counts.}
\usage{
add_chunks(frame, count, length = 2000)
}
\arguments{
\item{frame}{A set of counts}

\item{count}{The field containing wordcounts}

\item{length}{The desired chunk size. If smaller than 500 or so,
this will probably break against the existing pages.}
}
\value{
The input frame with a new column 'chunk'.
}
\description{
Add a set of 'chunks' to a tibble with page-level counts.
}
