% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hathidy.R
\name{grouped_tfidf}
\alias{grouped_tfidf}
\title{Add TF-IDF summary based on groupings.}
\usage{
grouped_tfidf(frame, word, count)
}
\arguments{
\item{frame}{A grouped data from}

\item{word}{The unquoted variable name storing terms for frequency}

\item{count}{The unquoted variable storing the count}
}
\value{
A data_frame with a column tfidf added.
}
\description{
The TF-IDF function in tidytext requires an explicit 'doc'
parameter; this applies it on the existing dataset groups.
}
