% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hathidy.R
\docType{package}
\name{hathidy}
\alias{hathidy}
\title{Hathidy: Work with wordcount data for 15 million books.}
\description{
In actual research, you shouldn't download files multile times, even if you work in multiple files. This package
uses a standard location.
}
