% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hathidy.R
\name{hathi_counts}
\alias{hathi_counts}
\title{Return Hathi Trust Extended Feature counts.}
\usage{
hathi_counts(
  htid,
  cols = c("page", "token"),
  sections = NULL,
  metadata = c("id"),
  cache = "feather",
  path = FALSE
)
}
\arguments{
\item{htid}{A Hathi Trust volume identifier, or a list of several.}

\item{cols}{The level of aggregation to return. Possible values: "page" (sequence in book), "token" (word), "POS"
(tagged part of speech), and "section." Default is c("page", "token"). If section is not requested, values are returned
only for the book's body (not headers and footers.)}

\item{sections}{Hathi divides books into 'header', 'footer', and 'body.' If you request 'section' as one of your
cols, all will be returned; but if you do not, only the counts for the body will be returned unless you manually
specificy, for example, (sections = c("header", "footer", "body"))}

\item{metadata}{Supplements the returned frame with volume-level metadata.
 Possible values are:  c("schemaVersion", "dateCreated", "volumeIdentifier", "accessProfile",
"rightsAttributes", "hathitrustRecordNumber", "enumerationChronology",
"sourceInstitution", "sourceInstitutionRecordNumber", "oclc", "isbn",
"issn", "lccn", "title", "imprint", "lastUpdateDate", "governmentDocument",
"pubDate", "pubPlace", "language", "bibliographicFormat", "genre", "issuance",
"typeOfResource", "classification", "names", "htBibUrl", "handleUrl")}

\item{cache}{Store a copy of the data for fast access next time. Default format is "csv".
"Support planned for 'parquet'.}

\item{path}{A direct filepath to a json dump. Use this if you are not using a stubbytree to store files.
If this is entered, 'htid' is ignored.}
}
\value{
a tibble, with columns created by the call.
}
\description{
Return Hathi Trust Extended Feature counts.
}
