% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hathidy.R
\name{hathi_counts}
\alias{hathi_counts}
\title{Return Hathi Trust Extended Feature counts.}
\usage{
hathi_counts(htid, cols = c("page", "token"), metadata = c())
}
\arguments{
\item{htid}{A Hathi Trust volume identifier, or a list of several.}

\item{cols}{The level of aggregation to return. Possible values: "page" (sequence in book), "token" (word), "POS"
(tagged part of speech), and "section." Default is c("page", "token"). If section is not requested, values are returned
only for the books body.}

\item{metadata}{Supplements the returned frame with volume-level metadata.
 Possible values are:  c("schemaVersion", "dateCreated", "volumeIdentifier", "accessProfile",
"rightsAttributes", "hathitrustRecordNumber", "enumerationChronology",
"sourceInstitution", "sourceInstitutionRecordNumber", "oclc", "isbn",
"issn", "lccn", "title", "imprint", "lastUpdateDate", "governmentDocument",
"pubDate", "pubPlace", "language", "bibliographicFormat", "genre", "issuance",
"typeOfResource", "classification", "names", "htBibUrl", "handleUrl")}
}
\value{
a tibble, with columns created by the call.
}
\description{
Return Hathi Trust Extended Feature counts.
}
