% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hathidy.R
\docType{package}
\name{hathidy}
\alias{hathidy}
\title{Hathidy: Work with wordcount data for 17 million books from the Hathi Trust}
\description{
An R interface for transparently downloading and analyzing page-level wordcount
data from the HathiTrust for over 17 million books.
}
